app/requirements.txt
flask==2.0.3
pywebview==3.6.2
wheel==0.37.1
pycairo==1.21.0
PyQt5==5.15.6
PyQtWebEngine==5.15.5
PyGObject==3.42.0
PySide2==5.15.2.1
QtPy==2.0.1
